Move token storage to the keychain and provide lockable tokens
This involves a reworking of several aspects and, most importantly,
the establishment of a test suite to ensure proper migration.

First, we create a new class (KeyringStore) which provides an interface
for marshalling/unmarshalling objects to the Keystore.

Second, we distinguish between OTP implementation itself (as defined
by RFC 4226) and the token metadata (as defined by the Google
Authenticator Key URI). These form the OTP and Token classes,
respectively.

Third, we migrate each of these classes to support (un)marshalling via
the iOS NSCoding interface. This allows persistence without having to
manually encode everything to a URI.

Fourth, we make it possible to lock just the OTP instances. This allows
token metadata to be read/written without unlocking. However, the core
secret, as a member of the OTP class, must be unlocked before accessing.

Fifth, we make it possible to create locked tokens and to lock or unlock
existing tokens (with appropriate authentication).

Last, we establish a test suite which includes tests for bugs we found
as well as testing migration from the old storage to the keyring.